Move WorkspaceVariableDisplayType to ark
